requirements.txt
requests==2.31.0
rich==13.7.0
httpx==0.27.0
typer==0.9.0
asyncio==3.4.3
aiohttp==3.9.3
python-dotenv==1.0.1 